# api-documentation.yaml
openapi: 3.0.3
info:
  title: Organization Service API
  description: |
    API documentation for the Organization Service which manages organizations, patients, and users.
    
    ## Authentication
    All endpoints (except /health) require Bearer token authentication.
    Header: `Authorization: Bearer <your-token>`
    
    ## For SUPER_ADMIN
    Add `X-Organization-ID` header to specify target organization.
    
    ## For ORG_ADMIN
    Uses organization from token automatically. Do NOT send X-Organization-ID header.
  version: 1.0.0
  contact:
    name: WailSalutem Health Care
    email: support@wailsalutem.com
servers:
  - url: http://localhost:8080
    description: Local development
  - url: https://api.wailsalutem.com
    description: Production
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    OrganizationID:
      name: X-Organization-ID
      in: header
      required: false
      description: Organization ID (required for SUPER_ADMIN, forbidden for ORG_ADMIN)
      schema:
        type: string
        format: uuid
  schemas:
...
    Patient:
      type: object
      properties:
        id:
          type: string
          format: uuid
        keycloak_user_id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone_number:
          type: string
        date_of_birth:
          type: string
          format: date
        address:
          type: string
        emergency_contact_name:
          type: string
        emergency_contact_phone:
          type: string
        medical_notes:
          type: string
        careplan_type:
          type: string
        careplan_frequency:
          type: string
        is_active:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
          nullable: true
    CreatePatientRequest:
      type: object
      required:
        - username
        - firstName
        - lastName
        - email
        - dateOfBirth
        - address
      properties:
        username:
          type: string
          example: john.patient
        temporaryPassword:
          type: string
          example: TempPass123!
        sendResetEmail:
          type: boolean
          default: false
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        email:
          type: string
          format: email
          example: john.doe@example.com
        phoneNumber:
          type: string
          example: "+1234567890"
        dateOfBirth:
          type: string
          format: date
          example: "1980-05-15"
        address:
          type: string
          example: 789 Patient St, City, State 12345
        emergencyContactName:
          type: string
          example: Jane Doe
        emergencyContactPhone:
          type: string
          example: "+1987654321"
        medicalNotes:
          type: string
          example: Diabetic, requires regular insulin
        careplanType:
          type: string
          example: intensive
        careplanFrequency:
          type: string
          example: daily
    UpdatePatientRequest:
      type: object
      properties:
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
        phone_number:
          type: string
        date_of_birth:
          type: string
          format: date
        address:
          type: string
        emergency_contact_name:
          type: string
        emergency_contact_phone:
          type: string
        medical_notes:
          type: string
        is_active:
          type: boolean
        careplan_type:
          type: string
        careplan_frequency:
          type: string
...
paths:
...
  /organization/patients:
    post:
      tags:
        - Patients
      summary: Create patient
      description: Create a new patient user
      operationId: createPatient
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientRequest'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

    get:
      tags:
        - Patients
      summary: List patients
      description: List all patients in organization
      operationId: listPatients
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/OrganizationID'
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Patient'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /organization/patients/{id}:
    get:
      tags:
        - Patients
      summary: Get patient by ID
      description: Get patient details
      operationId: getPatient
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationID'
      responses:
        '200':
          description: Patient details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found

    put:
      tags:
        - Patients
      summary: Update patient (full)
      description: Update patient information
      operationId: updatePatientPut
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientRequest'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found

    patch:
      tags:
        - Patients
      summary: Update patient (partial)
      description: Update patient information
      operationId: updatePatientPatch
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientRequest'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          description: Bad request
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found

    delete:
      tags:
        - Patients
      summary: Delete patient
      description: Soft delete patient
      operationId: deletePatient
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/OrganizationID'
      responses:
        '204':
          description: Patient deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
...